Union

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -3,9 +3,9 @@
 from typing import Union, List
 
 
-def sum_mixed_list(mxd_list: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_list: List[Union[int, float]]) -> float:
     """Calculates the sum of floats in the given list"""
-    sum: float = 0
+    sum: float = 0.00
     for num in mxd_list:
         sum += num
